# web/public/openapi.yaml
openapi: 3.0.3
info:
  title: GoRTMS - Go Real-Time Messaging System
  description: |
    GoRTMS is a real-time messaging system built with Go, featuring:
    - Domain-based message organization with queues
    - Consumer groups with offset management
    - Message routing with conditional rules
    - Dual authentication (JWT for web, HMAC for services)
    - Real-time monitoring and statistics
    - Runtime configuration management
  version: 1.0.0
  contact:
    name: GoRTMS API Support
    url: https://github.com/ajkula/GoRTMS
  license:
    name: Apache License
    url: http://www.apache.org/licenses/
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://your.own.gortms.server.com
    description: Production server
...
security:
  - bearerAuth: []
  - hmacAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user with Username/Password and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [Username, Password]
              properties:
                Username:
                  type: string
                  example: admin
                Password:
                  type: string
                  example: admin
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/bootstrap:
    post:
      tags: [Authentication]
      summary: Bootstrap admin user
      description: Create initial admin user when no users exist
      security: []
      responses:
        '201':
          description: Admin user created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  Password:
                    type: string
                    description: Generated admin Password
                    example: "aB3$mK9@pL2x"
        '409':
          description: Users already exist
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login

    hmacAuth:
      type: apiKey
      description: |
        HMAC authentication for service-to-service communication.
        Required headers:
        - X-Service-ID: Service account identifier
        - X-Timestamp: ISO 8601 timestamp (within 5-minute window)
        - X-Signature: HMAC-SHA256 signature of canonical request
        
        Canonical request format: `METHOD\nPATH\nBODY\nTIMESTAMP`
        Signature format: `sha256=<hex-encoded-hmac>`
      in: header
      name: X-Signature
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "3562599b-791e-4c8e-8a5b-89bbb98909a7"
        Username:
          type: string
          example: "admin"
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
          example: "2025-06-12T10:22:29Z"
        lastLogin:
          type: string
          format: date-time
          example: "2025-06-17T09:41:28Z"
        enabled:
          type: boolean
          example: true
    UserRole:
      type: string
      enum: [admin, user]
      example: admin

    # Service Accounts
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "bad_request"
              message:
                type: string
                example: "Invalid request body"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "unauthorized"
              message:
                type: string
                example: "Authentication required"
...